requirements.txt
numpy
pandas
python-dotenv
matplotlib
seaborn
plotly
streamlit
pydantic

ipykernel
jupyter

langchain
langchain-core
langchain-community
langchain-ollama
langchain-groq
langchain-openai
langchain-google-genai
langchain-huggingface
langgraph
langgraph.checkpoint.sqlite
langchain-unstructured
langchain-tavily

ollama
groq
openai
google-generativeai
faiss-cpu
pypdf
pypdf2
unstructured
docx
python-docx
wikipedia
youtube-search
google-api-python-client
arxiv
unstructured
unstructured-client
tavily-python
docx2txt

autogen
pyautogen

smolagents
smolagents[litellm]
transformers

exception

PyMuPDF
spacy
sentence-transformers
torch
scikit-learn

flask
flask-wtf
email-validator
flask-sqlalchemy
Flask-Session
